Friends done!!

+++ src/Post.jsx
@@ -3,24 +3,30 @@ import { connect } from "react-redux";
 import { Link, Redirect } from "react-router-dom";
 import Likes from "./Likes.jsx";
 import Comments from "./Comments.jsx";
+import Friends from "./Friends.jsx";
+import Requests from "./Requests.jsx";
 
 class UnconnectedPost extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      userProfile: {}
+      userProfile: {
+        friends: []
+      }
     };
   }
 
   componentDidMount = async () => {
+    this.getProfile();
+  };
+  getProfile = async () => {
     let data = new FormData();
     data.append("name", this.props.post.username);
     let response = await fetch("/getProfile", { method: "POST", body: data });
     let body = await response.text();
     let profile = JSON.parse(body);
-    console.log(body);
+    console.log(profile);
     this.setState({
-      ...this.state,
       userProfile: profile.user
     });
   };
@@ -53,38 +59,59 @@ class UnconnectedPost extends Component {
     }
     return false;
   };
+
+  // addFriend = () => {};
+
   render = () => {
+    console.log("this.state", this.state);
     let dispToggle = true;
     if (this.props.post.frontendPath === undefined) {
       dispToggle = false;
     }
-    //liked (postid+userId)
-    //post: if it exists in the collection. if it doesnt exist, you likeit
-    return (
-      <div style={{ border: "3px solid black", margin: "10px" }}>
-        <button onClick={this.deleteOne}>Delete this post </button>
-        <button onClick={this.delete}>Delete all posts </button>
-        <input type="text" onChange={this.editPostOnChange} />
-        <button onClick={this.editPost}>Edit your post </button>
-        <Link to={"/posts"}></Link>
-        <div className="post">{this.props.post.description}</div>
-        <div className="post">Posted by: {this.props.post.username}</div>
-        {this.isThereProfilePic() ? (
-          <Link to={"/users/" + this.state.userProfile.username}>
+    let showPost =
+      this.props.friends.includes(this.state.userProfile.username) ||
+      this.state.userProfile.username === this.props.username;
+    if (showPost) {
+      return (
+        <div style={{ border: "3px solid black", margin: "10px" }}>
+          <button onClick={this.deleteOne}>Delete this post </button>
+          <button onClick={this.delete}>Delete all posts </button>
+          <input type="text" onChange={this.editPostOnChange} />
+          <button onClick={this.editPost}>Edit your post </button>
+          <Link to={"/posts"}></Link>
+          <div className="post">{this.props.post.description}</div>
+          <div className="post">Posted by: {this.props.post.username}</div>
+          {this.isThereProfilePic() ? (
+            <Link to={"/users/" + this.state.userProfile.username}>
+              <img src={this.state.userProfile.profilePic} height="100px" />
+            </Link>
+          ) : (
+            <Link to={"/users/" + this.state.userProfile.username}>
+              <img src="/noimage.png" height="100px" />
+            </Link>
+          )}
+          {dispToggle ? (
+            <img src={this.props.post.frontendPath} height="400px" />
+          ) : (
+            ""
+          )}
+          <Likes post={this.props.post} />
+          <Comments post={this.props.post} />
+        </div>
+      );
+    } else
+      return (
+        <div>
+          {" "}
+          {this.state.userProfile.profilePic ? (
             <img src={this.state.userProfile.profilePic} height="100px" />
-          </Link>
-        ) : (
-          ""
-        )}
-        {dispToggle ? (
-          <img src={this.props.post.frontendPath} height="400px" />
-        ) : (
-          ""
-        )}
-        <Likes post={this.props.post} />
-        <Comments post={this.props.post} />
-      </div>
-    );
+          ) : (
+            <img src="/noimage.png" height="100px" />
+          )}{" "}
+          <Requests user={this.state.userProfile.username} />{" "}
+          {/* <button onClick={this.addFriend}>Add Friend</button>{" "} */}
+        </div>
+      );
   };
 }
 
@@ -92,7 +119,8 @@ let mapStateToProps = state => {
   return {
     description: state.description,
     file: state.file,
-    username: state.username
+    username: state.username,
+    friends: state.friends
   };
 };
